+++ src/lib/config.ts
@@ -1,6 +1,6 @@
 import { PUBLIC_DEBUG } from '$env/static/public';
 
-export const APP_VERSION = '2.2.1';
+export const APP_VERSION = '2.4.0';
 export const DEBUG = PUBLIC_DEBUG === 'true';
 
 export const checkAppVersion = (): boolean => {